+++ tools/web.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+
+    <style>
+    </style>
+</head>
+<body>
+    <div id="messages" class="alert alert-danger">
+
+    </div>
+    <button class="form-control" id="button_convert">konvertieren</button>
+    <div class="row">
+        <div class="col-md-6 col-sm-12">
+            <div class="input-group">
+                <label>diag_log getUnitLoadout player;</label>
+                <textarea class="form-control" cols="80" rows="10" id="input_unitLoadout"></textarea>
+            </div>
+        </div>
+        <div class="col-md-6 col-sm-12">
+            <div class="input-group">
+            <label>grad-loadout</label>
+                <textarea class="form-control" cols="80" rows="20" id="output_grad-loadout"></textarea>
+            </div>
+        </div>
+        <script>
+            exports = {};
+        </script>
+        <script src="unitLoadoutToGradLoadout.js"></script>
+
+        <script>
+            document.querySelector('#button_convert').addEventListener('click', function () {
+                try {
+                    var input = document.querySelector('#input_unitLoadout').value;
+                    var json = JSON.parse(input);
+                    document.querySelector('#output_grad-loadout').value = unitLoadoutToGradLoadout(json);
+                } catch (e) {
+                    document.querySelector('#messages').innerHTML += ('<br/>' + e.message);
+                }
+            });
+        </script>
+    </div>
+</body>
+</html>
